server/requirements.txt
flask
flask-sqlalchemy
uuid
flask-cors
python-dotenv
# opencv-python
# pillow
# numpy
# deepface